% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_scatter_themes}
\alias{get_scatter_themes}
\title{Generate a list with custom themes for scatter (y ~ x) plots}
\usage{
get_scatter_themes()
}
\value{
A list containing ggplot2 themes
}
\description{
Generate a list with custom themes for scatter (y ~ x) plots
}

