% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatting.R
\name{matrix_to_tbl_df}
\alias{matrix_to_tbl_df}
\title{Reformat a matrix int oa tbl_df table}
\usage{
matrix_to_tbl_df(a_matrix, row_name, col_name, value_name)
}
\arguments{
\item{a_matrix}{a matrix}

\item{row_name}{name of output column where row names of \code{a_matrix} will be stored}

\item{col_name}{name of output column where column names of \code{a_matrix} will be stored}

\item{col_name}{name of output column where values from \code{a_matrix} will be stored}
}
\value{
a tbl_df object with columns matching the row names, column names and values of \code{a_matrix}
}
\description{
Reformat a matrix int oa tbl_df table
}
\examples{
input_matrix <- matrix(1:9, ncol = 3)
rownames(input_matrix) <- letters[1:3]
colnames(input_matrix) <- LETTERS[1:3]
matrix_to_tbl_df(input_matrix, "row", "col", "val")
}
\seealso{
\code{\link{tbl_df_to_matrix}}
}

